create a new usd.py
